fix(live-preview-react): properly types useLivePreview hook
