+++ packages/live-preview-react/src/index.ts
@@ -1,3 +1,5 @@
+import type { MergeLiveDataArgs } from '@payloadcms/live-preview/types'
+
 import { ready, subscribe, unsubscribe } from '@payloadcms/live-preview'
 import { useCallback, useEffect, useRef, useState } from 'react'
 
@@ -6,11 +8,9 @@ import { useCallback, useEffect, useRef, useState } from 'react'
 // To prevent the flicker of stale data while the post message is being sent,
 // you can conditionally render loading UI based on the `isLoading` state
 
-export const useLivePreview = <T extends any>(props: {
-  depth?: number
-  initialData: T
-  serverURL: string
-}): {
+export const useLivePreview = <T extends any>(
+  props: Omit<MergeLiveDataArgs, 'fieldSchema' | 'incomingData' | 'returnNumberOfRequests'>,
+): {
   data: T
   isLoading: boolean
 } => {
