+++ packages/live-preview-react/tsconfig.json
@@ -8,6 +8,9 @@
     "rootDir": "./src" /* Specify the root folder within your source files. */,
     "jsx": "react"
   },
+  "paths": {
+    "@payloadcms/live-preview/types": ["../live-preview/types"]
+  },
   "exclude": [
     "dist",
     "build",
