+++ packages/live-preview/src/mergeData.ts
@@ -1,16 +1,6 @@
-import type { fieldSchemaToJSON } from 'payload/utilities'
+import type { MergeLiveDataArgs } from './types'
 
 import { traverseFields } from './traverseFields'
-
-export type MergeLiveDataArgs<T> = {
-  apiRoute?: string
-  depth: number
-  fieldSchema: ReturnType<typeof fieldSchemaToJSON>
-  incomingData: Partial<T>
-  initialData: T
-  returnNumberOfRequests?: boolean
-  serverURL: string
-}
 
 export const mergeData = async <T>({
   apiRoute,
